first pass at removing eif code. need to replace some code for testing that was creating stubs using the eif_ingestor still
