+++ src/publisher/tests/base.py
@@ -1,9 +1,57 @@
 from io import StringIO
-import os, json
+import os, json, copy
 from django.test import TestCase as DjangoTestCase, TransactionTestCase
 from publisher import models, utils
 from django.core.management import call_command
 import unittest
+
+'''
+def filldict(ddict, keys, default):
+    def filldictslot(ddict, key, val):
+        if key not in ddict:
+            ddict[key] = val
+    data = copy.deepcopy(ddict)
+    for key in keys:
+        if isinstance(key, tuple):
+            key, val = key
+        else:
+            val = default
+        filldictslot(data, key, val)
+    return data
+
+def add_or_update_article(**article_data):
+    """TESTING ONLY. given article data it attempts to find the
+    article and update it, otherwise it will create it, filling
+    any missing keys with dummy data. returns the created article."""
+    assert 'doi' in article_data or 'manuscript_id' in article_data, \
+        "a value for 'doi' or 'manuscript_id' *must* exist"
+
+    if 'manuscript_id' in article_data:
+        article_data['doi'] = utils.msid2doi(article_data['manuscript_id'])
+    elif 'doi' in article_data:
+        article_data['manuscript_id'] = utils.doi2msid(article_data['doi'])
+
+    filler = [
+        'title',
+        'doi',
+        'manuscript_id',
+        ('volume', 1),
+        'path',
+        'article-type',
+        ('ejp_type', 'RA'),
+        ('version', 1),
+        ('pub-date', '2012-01-01'),
+        ('status', 'vor'),
+    ]
+    article_data = utils.filldict(article_data, filler, 'pants-party')
+    # return eif_ingestor.import_article(journal(), article_data, create=True, update=True)
+    return article_data
+
+'''
+
+#
+#
+#
 
 class SimpleBaseCase(unittest.TestCase):
     this_dir = os.path.dirname(os.path.realpath(__file__))
