--- src/publisher/tests/fixtures/ppp/00353.1/elife-00353-v1.json
@@ -1,1 +1,0 @@
-{"status": "VOR", "article-type": "discussion", "contributors": [{"surname": "Marder", "corresp": true, "affiliations": [{"dept": "Department of Biology and the Volen National Center for Complex Systems", "city": "Waltham", "email": "marder@brandeis.edu", "institution": "Brandeis University", "country": "United States"}], "given-names": "Eve", "role": "Senior Editor", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-1021"}], "pub-date": "2012-12-13T00:00:00Z", "title": "A good life", "article-id": "00353", "referenced": {"competing-interest": {"conf1": "The author declares that no competing interests exist"}}, "update": "2012-12-13T00:00:00Z", "publish": true, "volume": 1, "article-version-id": "00353.1", "version": 1, "keywords": {"author-keywords": ["careers in science", "grad school", "postdoc"]}, "path": "content/1/e00353v1", "elocation-id": "e00353", "categories": {"sub-display-channel": ["Living Science"], "display-channel": ["Feature Article"]}, "doi": "10.7554/eLife.00353"}
