--- src/publisher/tests/fixtures/ppp/00385.1/elife-00385-v1.json
@@ -1,1 +1,0 @@
-{"status": "VOR", "article-type": "article-commentary", "contributors": [{"surname": "Jha", "corresp": true, "affiliations": [{"dept": "Dalla Lana School of Public Health, and the Centre for Global Health Research", "email": "jhap@smh.ca", "institution": "University of Toronto", "country": "Canada"}], "given-names": "Prabhat", "role": "Senior Editor", "references": {"affiliation": ["aff1"], "competing-interest": ["conf1"]}, "type": "author", "id": "author-1017"}], "pub-date": "2012-12-18T00:00:00Z", "title": "Mathematics and malaria", "article-id": "00385", "related-articles": [{"href": "10.7554/eLife.00093", "type": "commentary-article"}], "referenced": {"affiliation": {"aff1": {"dept": "Dalla Lana School of Public Health, and the Centre for Global Health Research", "country": "Canada", "email": "jhap@smh.ca", "institution": "University of Toronto"}}, "competing-interest": {"conf1": "The author declares that no competing interests exist"}}, "update": "2012-12-18T00:00:00Z", "publish": true, "article-version-id": "00385.1", "volume": 1, "citations": {"bib11": {"source": "N Engl J Med", "authors": [{"group-type": "author", "collab": "The RTS,S clinical trials partnership"}], "doi": "10.1056/NEJMoa1208394", "title": "A phase 3 trial of RTS,S/AS01 malaria vaccine in African infants", "year": "2012"}, "bib10": {"source": "Adv Parasitol", "authors": [{"surname": "Snow", "given-names": "RW", "group-type": "author"}, {"surname": "Amratia", "given-names": "P", "group-type": "author"}, {"surname": "Kabaria", "given-names": "CW", "group-type": "author"}, {"surname": "Noor", "given-names": "AM", "group-type": "author"}, {"surname": "Marsh", "given-names": "K", "group-type": "author"}], "doi": "10.1016/B978-0-12-394303-3.00010-4", "title": "The changing limits and incidence of malaria in Africa: 1939\u20132009", "year": "2012"}, "bib9": {"source": "Proc Roy Soc A", "authors": [{"surname": "Ross", "given-names": "R", "group-type": "author"}], "doi": "10.1098/rspa.1916.0007", "title": "An application of the theory of probabilities to the study of a priori pathometry. Part I", "year": "1916"}, "bib8": {"source": "Adv Parasitol", "authors": [{"surname": "McKenzie", "given-names": "FE", "group-type": "author"}, {"surname": "Smith", "given-names": "DL", "group-type": "author"}, {"surname": "O'Meara", "given-names": "WP", "group-type": "author"}, {"surname": "Riley", "given-names": "EM", "group-type": "author"}], "doi": "10.1016/S0065-308X(08)00201-7", "title": "Strain theory of malaria: the first 50 years", "year": "2008"}, "bib1": {"source": "eLife", "authors": [{"surname": "Artzy-Randrup", "given-names": "Y", "group-type": "author"}, {"surname": "Rorick", "given-names": "M", "group-type": "author"}, {"surname": "Day", "given-names": "K", "group-type": "author"}, {"surname": "Chen", "given-names": "D", "group-type": "author"}, {"surname": "Dobson", "given-names": "A", "group-type": "author"}, {"surname": "Pascual", "given-names": "M", "group-type": "author"}], "doi": "10.7554/eLife.00093", "title": "Population structuring of multi-copy, antigen-encoding genes in Plasmodium falciparum", "year": "2012"}, "bib3": {"source": "Disease control priorities in developing countries", "authors": [{"surname": "Breman", "given-names": "JG", "group-type": "author"}, {"surname": "Mills", "given-names": "A", "group-type": "author"}, {"surname": "Snow", "given-names": "RW", "group-type": "author"}, {"surname": "Mulligan", "given-names": "JA", "group-type": "author"}, {"surname": "Lengeler", "given-names": "C", "group-type": "author"}, {"surname": "Mendis", "given-names": "K", "group-type": "author"}, {"surname": "Jamison", "given-names": "DT", "group-type": "editor"}, {"surname": "Breman", "given-names": "JG", "group-type": "editor"}, {"surname": "Measham", "given-names": "AR", "group-type": "editor"}, {"surname": "Alleyne", "given-names": "G", "group-type": "editor"}, {"surname": "Claeson", "given-names": "M", "group-type": "editor"}, {"surname": "Evans", "given-names": "DB", "group-type": "editor"}, {"surname": "Jha", "given-names": "P", "group-type": "editor"}, {"surname": "Mills", "given-names": "A", "group-type": "editor"}, {"surname": "Musgrove", "given-names": "P", "group-type": "editor"}], "title": "Chapter 21: Conquering malaria", "year": "2006"}, "bib2": {"source": "Inf Sci (Ny)", "authors": [{"surname": "Bauer", "given-names": "AL", "group-type": "author"}, {"surname": "Beauchemin", "given-names": "CA", "group-type": "author"}, {"surname": "Perelsond", "given-names": "AS", "group-type": "author"}], "title": "Agent-based modeling of host\u2013pathogen systems: the successes and challenges", "year": "2009"}, "bib5": {"source": "Lancet", "authors": [{"surname": "Dhingra", "given-names": "N", "group-type": "author"}, {"surname": "Jha", "given-names": "P", "group-type": "author"}, {"surname": "Sharma", "given-names": "VP", "group-type": "author"}, {"surname": "Cohen", "given-names": "AA", "group-type": "author"}, {"surname": "Jotkar", "given-names": "RM", "group-type": "author"}, {"surname": "Rodriguez", "given-names": "PS", "group-type": "author"}], "doi": "10.1016/S0140-6736(10)60831-8", "title": "Adult and child malaria mortality in India: a nationally representative mortality survey", "year": "2010"}, "bib4": {"source": "Developmental Psychology", "authors": [{"surname": "Brown", "given-names": "BB", "group-type": "author"}, {"surname": "Clasen", "given-names": "DR", "group-type": "author"}, {"surname": "Eicher", "given-names": "SA", "group-type": "author"}], "doi": "10.1037/0012-1649.22.4.521", "title": "Perceptions of peer pressure, peer conformity dispositions, and self-reported behavior among adolescents", "year": "1986"}, "bib7": {"source": "Nat Med", "authors": [{"surname": "Gupta", "given-names": "S", "group-type": "author"}, {"surname": "Maiden", "given-names": "MC", "group-type": "author"}, {"surname": "Feavers", "given-names": "IM", "group-type": "author"}, {"surname": "Nee", "given-names": "S", "group-type": "author"}, {"surname": "May", "given-names": "RM", "group-type": "author"}, {"surname": "Anderson", "given-names": "RM", "group-type": "author"}], "doi": "10.1038/nm0496-437", "title": "The maintenance of strain structure in populations of recombining infectious agents", "year": "1996"}, "bib6": {"source": "Science", "authors": [{"surname": "Gupta", "given-names": "S", "group-type": "author"}, {"surname": "Trenholme", "given-names": "K", "group-type": "author"}, {"surname": "Anderson", "given-names": "RM", "group-type": "author"}, {"surname": "Day", "given-names": "KP", "group-type": "author"}], "doi": "10.1126/science.8310293", "title": "Antigenic diversity and the transmission dynamics of Plasmodium falciparum", "year": "1994"}}, "version": 1, "keywords": {"author-keywords": ["parasite population structure", "epidemiological dynamic", "antigenic diversity", "immune selection", "Plasmodium falciparum", "var multi-gene family"], "research-organism": ["Other"]}, "path": "content/1/e00385v1", "elocation-id": "e00385", "categories": {"sub-display-channel": ["Infectious Diseases"], "heading": ["Microbiology and Infectious Disease"], "display-channel": ["Insight"]}, "doi": "10.7554/eLife.00385"}
