--- src/publisher/tests/fixtures/ppp/01328.1/elife-01328-v1.json
@@ -1,1 +1,0 @@
-{"status": "VOR", "article-type": "correction", "contributors": [{"surname": "Guo", "given-names": "Yusong", "type": "author", "id": "author-1916", "references": {"competing-interest": ["conf2"]}}, {"surname": "Zanetti", "given-names": "Giulia", "type": "author", "id": "author-2244", "references": {"competing-interest": ["conf2"]}}, {"surname": "Schekman", "given-names": "Randy", "type": "author", "id": "author-1032", "references": {"competing-interest": ["conf1"]}}], "pub-date": "2013-08-28T00:00:00Z", "title": "Correction: A novel GTP-binding protein\u2013adaptor protein complex responsible for export of Vangl2 from the <italic>trans</italic> Golgi network", "article-id": "01328", "related-articles": [{"href": "10.7554/eLife.00160", "type": "corrected-article"}], "update": "2013-08-28T00:00:00Z", "publish": true, "volume": 2, "article-version-id": "01328.1", "version": 1, "referenced": {"competing-interest": {"conf2": "The other authors have declared that no competing interests exist.", "conf1": "RS: Editor-in-Chief, <italic>eLife</italic>."}}, "path": "content/2/e01328v1", "elocation-id": "e01328", "categories": {"heading": ["Biochemistry", "Cell Biology"], "display-channel": ["Correction"]}, "doi": "10.7554/eLife.01328"}
