--- src/publisher/tests/fixtures/ppp/02619.1/elife-02619-v1.json
@@ -1,1 +1,0 @@
-{"status": "VOR", "article-type": "editorial", "contributors": [{"surname": "Harrison", "corresp": true, "affiliations": [{"country": "United Kingdom", "email": "production@elifesciences.org", "institution": "<italic>eLife</italic>", "city": "Cambridge"}], "given-names": "Melissa", "references": {"affiliation": ["aff1"], "competing-interest": ["conf1"]}, "orcid": "http://orcid.org/0000-0003-3523-4408", "type": "author", "id": "author-1004"}], "pub-date": "2014-04-29T00:00:00Z", "title": "Coming soon to a screen near you", "article-id": "02619", "referenced": {"affiliation": {"aff1": {"country": "United Kingdom", "email": "production@elifesciences.org", "institution": "<italic>eLife</italic>", "city": "Cambridge"}}, "competing-interest": {"conf1": "The author declares that no competing interests exist."}}, "update": "2014-04-29T00:00:00Z", "publish": true, "volume": 3, "article-version-id": "02619.1", "version": 1, "keywords": {"author-keywords": ["scientific publishing", "open access", "eLife", "xml", "repositories"]}, "path": "content/3/e02619v1", "elocation-id": "e02619", "categories": {"sub-display-channel": ["Scientific Publishing"], "display-channel": ["Editorial"]}, "doi": "10.7554/eLife.02619"}
