--- src/publisher/tests/fixtures/ppp/03401.2/elife-03401-v2.json
@@ -1,1 +1,0 @@
-{"status": "POA", "article-type": "research-article", "contributors": [{"surname": "Esvelt", "corresp": true, "affiliations": [{"country": "United States", "email": "kevin.esvelt@wyss.harvard.edu", "institution": "Wyss Institute for Biologically Inspired Engineering, Harvard Medical School", "city": "Boston"}], "given-names": "Kevin M", "references": {"competing-interest": ["conf2"]}, "type": "author", "id": "author-16152"}, {"surname": "Smidler", "affiliations": [{"country": "United States", "institution": "Harvard School of Public Health", "city": "Boston"}], "given-names": "Andrea L", "references": {"competing-interest": ["conf3"]}, "type": "author", "id": "author-14374"}, {"surname": "Catteruccia", "affiliations": [{"country": "United States", "institution": "Harvard School of Public Health", "city": "Boston"}], "given-names": "Flaminia", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-14375"}, {"surname": "Church", "affiliations": [{"country": "United States", "institution": "Wyss Institute for Biologically Inspired Engineering, Harvard Medical School", "city": "Boston"}], "given-names": "George M", "references": {"competing-interest": ["conf4"]}, "type": "author", "id": "author-16153"}, {"surname": "Tautz", "affiliations": [{"country": "Germany", "institution": "Max Planck Institute for Evolutionary Biology"}], "given-names": "Diethard", "role": "Reviewing editor", "type": "editor", "id": "author-1191"}], "pub-date": "2014-07-16T00:00:00Z", "title": "Concerning RNA-guided gene drives for the alteration of wild populations", "article-id": "03401", "referenced": {"competing-interest": {"conf3": "Andrea L Smidler, Filed for a patent concerning RNA-guided gene drives.", "conf2": "Kevin M Esvelt, Filed for a patent concerning RNA-guided gene drives.", "conf1": "The other authors declare that no competing interests exist.", "conf4": "George M Church, Filed for a patent concerning RNA-guided gene drives."}}, "update": "2014-07-16T00:00:00Z", "publish": true, "volume": 3, "article-version-id": "03401.2", "version": 2, "keywords": {"research-organism": ["None"]}, "path": "content/3/e03401v2", "elocation-id": "e03401", "categories": {"heading": ["Ecology", "Genes and Chromosomes"], "display-channel": ["Feature Article"]}, "doi": "10.7554/eLife.03401"}
