--- src/publisher/tests/fixtures/ppp/06250.1/elife-06250-v1.json
@@ -1,1 +1,0 @@
-{"status": "POA", "article-type": "research-article", "contributors": [{"surname": "Johnston", "corresp": true, "affiliations": [{"dept": "School of Life Sciences", "city": "Brighton", "email": "J.johnston@sussex.ac.uk", "institution": "University of Sussex", "country": "United Kingdom"}], "given-names": "Jamie", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-23463"}, {"surname": "Lagnado", "affiliations": [{"dept": "School of Life Sciences", "city": "Brighton", "institution": "University of Sussex", "country": "United Kingdom"}], "given-names": "Leon", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-23801"}, {"surname": "Carandini", "affiliations": [{"country": "United Kingdom", "institution": "University College London"}], "given-names": "Matteo", "role": "Reviewing editor", "type": "editor", "id": "author-13133"}], "pub-date": "2015-03-18T00:00:00Z", "title": "General features of the retinal connectome determinethe computation of motion anticipation", "article-id": "06250", "referenced": {"competing-interest": {"conf1": "The authors declare that no competing interests exist."}}, "update": "2015-03-18T00:00:00Z", "publish": true, "volume": 4, "article-version-id": "06250.1", "version": 1, "keywords": {"research-organism": ["Other"]}, "path": "content/4/e06250v1", "elocation-id": "e06250", "categories": {"heading": ["Neuroscience"], "display-channel": ["Research Article"]}, "doi": "10.7554/eLife.06250"}
