--- src/publisher/tests/fixtures/ppp/08025.1/elife-08025-v1.json
@@ -1,1 +1,0 @@
-{"status": "POA", "article-type": "research-article", "contributors": [{"surname": "Kim", "affiliations": [{"dept": "Department of Ophthalmology and Visual Sciences", "city": "Saint Louis", "institution": "Washington University School of Medicine", "country": "United States"}], "given-names": "Tahnbee", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-29066"}, {"surname": "Soto", "affiliations": [{"dept": "Department of Ophthalmology and Visual Sciences", "city": "Saint Louis", "institution": "Washington University School of Medicine", "country": "United States"}], "given-names": "Florentina", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-29068"}, {"surname": "Kerschensteiner", "corresp": true, "affiliations": [{"dept": "Department of Ophthalmology and Visual Sciences", "city": "Saint Louis", "email": "dkerschensteiner@wustl.edu", "institution": "Washington University School of Medicine", "country": "United States"}], "given-names": "Daniel", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-29026"}, {"surname": "Borst", "affiliations": [{"country": "Germany", "institution": "Max Planck Institute of Neurobiology"}], "given-names": "Alexander", "role": "Reviewing editor", "type": "editor", "id": "author-6613"}], "pub-date": "2015-05-19T00:00:00Z", "title": "An excitatory amacrine cell detects object motion and provides feature-selective input to ganglion cells in the mouse retina", "article-id": "08025", "referenced": {"competing-interest": {"conf1": "The authors declare that no competing interests exist."}}, "update": "2015-05-19T00:00:00Z", "publish": true, "volume": 4, "article-version-id": "08025.1", "version": 1, "keywords": {"research-organism": ["Mouse"]}, "path": "content/4/e08025v1", "elocation-id": "e08025", "categories": {"heading": ["Neuroscience"], "display-channel": ["Short Report"]}, "doi": "10.7554/eLife.08025"}
