--- src/publisher/tests/fixtures/ppp/09571.1/elife-09571-v1.json
@@ -1,1 +1,0 @@
-{"status": "POA", "article-type": "research-article", "contributors": [{"surname": "Zylicz", "affiliations": [{"dept": "Wellcome Trust/Cancer Research UK Gurdon Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Jan J", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-36915"}, {"surname": "Dietmann", "affiliations": [{"dept": "Wellcome Trust/Medical Research Council Stem Cell Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Sabine", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-36916"}, {"surname": "G\u00fcnesdogan", "affiliations": [{"dept": "Wellcome Trust/Cancer Research UK Gurdon Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Ufuk", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-4081"}, {"surname": "Hackett", "affiliations": [{"dept": "Wellcome Trust/Cancer Research UK Gurdon Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Jamie A", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-36917"}, {"surname": "Cougot", "affiliations": [{"dept": "Wellcome Trust/Cancer Research UK Gurdon Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Delphine", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-36918"}, {"surname": "Lee", "affiliations": [{"dept": "Wellcome Trust/Cancer Research UK Gurdon Institute", "city": "Cambridge", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "Caroline", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-36920"}, {"surname": "Surani", "corresp": true, "affiliations": [{"dept": "Wellcome Trust / Cancer Research UK Gurdon Institute", "city": "Cambridge", "email": "a.surani@gurdon.cam.ac.uk", "institution": "University of Cambridge", "country": "United Kingdom"}], "given-names": "M Azim", "references": {"competing-interest": ["conf1"]}, "type": "author", "id": "author-17591"}, {"surname": "Akhtar", "affiliations": [{"country": "Germany", "institution": "Max Planck Institute for Immunobiology and Epigenetics"}], "given-names": "Asifa", "role": "Reviewing editor", "type": "editor", "id": "author-1035"}], "pub-date": "2015-11-09T00:00:00Z", "title": "Chromatin dynamics and the role of G9a in gene regulation and enhancer silencing during early mouse development", "article-id": "09571", "referenced": {"competing-interest": {"conf1": "The authors declare that no competing interests exist."}}, "update": "2015-11-09T00:00:00Z", "publish": true, "volume": 4, "article-version-id": "09571.1", "version": 1, "keywords": {"research-organism": ["Mouse"]}, "path": "content/4/e09571v1", "elocation-id": "e09571", "categories": {"heading": ["Developmental Biology and Stem Cells", "Genes and Chromosomes"], "display-channel": ["Research Article"]}, "doi": "10.7554/eLife.09571"}
